x/exchange/types/message_exchange_token_test.go
package types

import (
	"testing"

	"bu-chain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgExchangeToken_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgExchangeToken
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgExchangeToken{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgExchangeToken{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
